Fix "failed setting rlimit: invalid argument" on osx
Running on osx, the attempt to change the rlimit fails with "invalid argument".

It appears that the value of rlimit.Max is `9223372036854775807`, and that
trying to set to that value fails. More on this at
https://github.com/golang/go/issues/30401.

We fix this by setting the value to the somewhat magic 24576 when on osx (darwin).